Refactor: Move auxiliary functions to where they are used
Delete dataframe_function_handler.py

Move functions to modules which use them:
- astro.sql.operators.transform
- astro.sql.operators.sql_dataframe
