+++ src/astro/sql/operators/transform.py
@@ -1,18 +1,70 @@
 import inspect
-from typing import Callable, Dict, Iterable, Mapping, Optional, Union
+from typing import Callable, Dict, Iterable, Mapping, Optional, Tuple, Union
 
+import pandas as pd
 from airflow.decorators.base import DecoratedOperator, task_decorator_factory
 from airflow.exceptions import AirflowException
 from sqlalchemy.sql.functions import Function
 
 from astro.databases import create_database
 from astro.sql.table import Table
-from astro.utils.dataframe_function_handler import (
-    load_op_arg_dataframes_into_sql,
-    load_op_kwarg_dataframes_into_sql,
-)
 from astro.utils.sql_handler import handle_schema
 from astro.utils.table_handler_new import find_first_table
+
+
+def load_op_arg_dataframes_into_sql(
+    conn_id: str, op_args: Tuple, target_table: Table
+) -> Tuple:
+    """
+    Identifies dataframes in op_args and loads them to the table
+
+    :param conn_id:
+    :param op_args:
+    :param target_table:
+    :return:
+    """
+    final_args = []
+    database = create_database(conn_id=conn_id)
+    for arg in op_args:
+        if isinstance(arg, pd.DataFrame):
+            database.load_pandas_dataframe_to_table(
+                source_dataframe=arg, target_table=target_table
+            )
+            final_args.append(target_table)
+        elif isinstance(arg, Table):
+            arg = database.populate_table_metadata(arg)
+            final_args.append(arg)
+        else:
+            final_args.append(arg)
+    return tuple(final_args)
+
+
+def load_op_kwarg_dataframes_into_sql(
+    conn_id: str, op_kwargs: Dict, target_table: Table
+) -> Dict:
+    """
+    Identifies dataframes in op_kwargs and loads them to the table
+
+    :param conn_id:
+    :param op_kwargs:
+    :param target_table:
+    :return:
+    """
+    final_kwargs = {}
+    database = create_database(conn_id=conn_id)
+    for key, value in op_kwargs.items():
+        if isinstance(value, pd.DataFrame):
+            df_table = target_table.create_similar_table()
+            database.load_pandas_dataframe_to_table(
+                source_dataframe=value, target_table=df_table
+            )
+            final_kwargs[key] = df_table
+        elif isinstance(value, Table):
+            value = database.populate_table_metadata(value)
+            final_kwargs[key] = value
+        else:
+            final_kwargs[key] = value
+    return final_kwargs
 
 
 class TransformOperator(DecoratedOperator):
